Add files via upload
Customer reviews provide valuable insights into customer opinions and
perceptions, but traditional sentiment analysis methods can be limited in their
ability to identify key factors and corresponding sentiments. The goal is to
develop a solution that can provide a more comprehensive analysis of customer
reviews, taking into account the context of customer language. The
challenge is to find an optimized solution that can perform this analysis quickly
and accurately, given the large volume of customer reviews available.
 The aspect-based sentiment evaluation method provided here is an
integrated approach that combines area lexicons and rules for analyzing app
reviews. The method's purpose is to extract essential components from reviews
and categorize the sentiment associated with each factor. By using a
combination of lexicons and rules, the proposed approach aims to provide a
more accurate and nuanced analysis of the customer reviews. The system
processes the review data and determines the overall sentiment of each review
using the Nave Bayes and Artificial Bee Colony algorithms in conjunction with
the lexicons and rules.
 The ABC algorithm allows for an optimization of the analysis process,
ensuring that the sentiment analysis is performed quickly and accurately. The
integration of the lexicons and rules provides additional context and information
to the analysis, resulting in a more comprehensive and reliable analysis of the
customer reviews.